+++ sphinx/changelog.rst
@@ -2,7 +2,15 @@
 Changelog
 #########
 
+* `Version 1.0.4`_ (29.4.2016)
 * `Version 1.0.3`_ (31.3.2016)
+
+Version 1.0.4
+=============
+
+- (fbt_run) Include SKIPPED tests into results file with spec. outcome
+- (fbt_run) Check that Firebird is running before test execution
+- (fbt_analyze) Show time performance of tests
 
 Version 1.0.3
 =============